refactor(frontend): create form types and qiuze-form component bb-88
