+++ apps/frontend/src/libs/types/types.ts
@@ -1,5 +1,10 @@
 export { type AsyncThunkConfig } from "./async-thunk-config.type.js";
 export {
+	type Control,
+	type FieldPath,
+	type FieldValues,
+} from "./form.types.js";
+export {
 	type ServerErrorDetail,
 	type ServerErrorResponse,
 	type ValidationSchema,
