+++ apps/frontend/src/pages/quiz/libs/components/quiz-form/quiz-form.tsx
@@ -1,0 +1,20 @@
+import { QuizQuestion } from "~/libs/components/components.js";
+import { useAppForm } from "~/libs/hooks/hooks.js";
+
+type FormValues = {
+	value: string;
+};
+
+const QuizForm = () => {
+	const { control } = useAppForm<FormValues>({
+		defaultValues: { value: "" },
+	});
+
+	return (
+		<form>
+			<QuizQuestion control={control} />
+		</form>
+	);
+};
+
+export { QuizForm };
